로그인, 중복로그인, 아이디 비밀번호 체크 / 로그아웃(로그인 상태에서만 되게) / article write, modify, delete 시 로그인 체크/ modify, delete시 권한체크

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -19,6 +19,7 @@ import com.example.demo.vo.ResultData;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 @Controller // 프로젝트에서 얘는 컨트롤러라고 인식하게끔 하는것
 public class UsrArticleController {
@@ -30,8 +31,14 @@ public class UsrArticleController {
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
 	// ResultData<Article> : data1의 타입이 Article이라고 명시
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession httpSession, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후에 이용해주세요.");
+		}
 
+		int memberId = (int) httpSession.getAttribute("loginedMemberId");
+		
 		if (Ut.isEmptyOrNull(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
@@ -39,13 +46,13 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 
-		ResultData writeArticleRd = articleService.writeArticle(title, body);
+		ResultData writeArticleRd = articleService.writeArticle(memberId, title, body);
 
 		int id = (int) writeArticleRd.getData1();
 
 		Article article = articleService.getArticleById(id);
 
-		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
+		return ResultData.newData(writeArticleRd, article);
 	}
 
 	// 게시글 목록
@@ -74,12 +81,20 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Article> doDelete(int id) {
+	public ResultData<Article> doDelete(HttpSession httpSession, int id) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후에 이용해주세요.");
+		}
 
 		Article foundArticle = articleService.getArticleById(id);
 
 		if (foundArticle == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
+		}
+		
+		if (foundArticle.getMemberId() != (int) httpSession.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 대한 권한이 없습니다.", id));
 		}
 
 		articleService.deleteArticle(id);
@@ -91,12 +106,20 @@ public class UsrArticleController {
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
 	// object를 쓰는 이유는 어떨때는 return을 String으로 하고 어떨때는 Article 타입으로 하기 위해
-	public ResultData<Article> doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후에 이용해주세요.");
+		}
 
 		Article foundArticle = articleService.getArticleById(id);
 
 		if (foundArticle == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
+		}
+		
+		if (foundArticle.getMemberId() != (int) httpSession.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 대한 권한이 없습니다.", id));
 		}
 
 		articleService.modifyArticle(id, title, body);
